Rename MasterSlaveSpringBootStarterTest
