+++ shardingsphere-features/shardingsphere-master-slave/shardingsphere-master-slave-spring/shardingsphere-master-slave-spring-boot-starter/src/test/resources/application-masterslave.properties
@@ -15,6 +15,8 @@
 # limitations under the License.
 #
 
+spring.shardingsphere.rules.master-slave.load-balancers.random.type=RANDOM
 
-spring.shardingsphere.rules.master-slave.load-balancers.fixture.type=FIXTURE
-
+spring.shardingsphere.rules.master-slave.data-sources.ds_ms.master-data-source-name=ds_master
+spring.shardingsphere.rules.master-slave.data-sources.ds_ms.slave-data-source-names=ds_slave_0,ds_slave_1
+spring.shardingsphere.rules.master-slave.data-sources.ds_ms.load-balancer-name=random
